requirements.txt
Flask==3.0.0
twilio==8.10.0
google-generativeai==0.3.0
dspy-ai==2.3.0
APScheduler==3.10.4